corrections

+++ 0x03-python-data_structures/100-print_python_list_info.c
@@ -10,15 +10,15 @@
 {
     PyListObject *list;
     PyObject *item;
-    int size, i;
+    long int size, i;
 
     size = Py_SIZE(p);
-    printf("[*] Size of the Python List = %d\n", size);
+    printf("[*] Size of the Python List = %ld\n", size);
     list = (PyListObject *)p;
-    printf("[*] Allocated = %d\n", list->allocated)
+    printf("[*] Allocated = %ld\n", list->allocated)
     for (i = 0; i < size; i++)
     {
         item = PyList_GetItem(p, i);
-        printf("Element %d: %s\n", i, Py_TYPE(item)->tp_name)
+        printf("Element %ld: %s\n", i, Py_TYPE(item)->tp_name)
     }
 }
